Extracted UuidValidator from TokenModel
