+++ src/lib/uuid/UuidValidator.js
@@ -1,0 +1,8 @@
+export default class UuidValidator {
+  static isUuidValid(uuid) {
+    const regExp = new RegExp(
+      '^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$',
+    );
+    return regExp.test(uuid);
+  }
+}
